requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2

drf-spectacular~=0.28.0
drf-yasg~=1.21.8

pillow~=11.0.0
